requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
langchain-groq
sentence_transformers
groq
cassio
beautifulsoup4
tf-keras
ragas
nltk
peft
rapidfuzz
torch
transformers
pytest